+++ src/StashUnstashPlugin.groovy
@@ -2,6 +2,8 @@ public class StashUnstashPlugin implements Plugin, Resettable {
     public static final String DEFAULT_STASH_NAME = 'buildArtifact'
     private static String pattern
     private static String patternFile
+    private static String unstashVariableName
+    private String stashedFilename
 
     public static withArtifact(String pattern) {
         this.pattern = pattern
@@ -10,6 +12,11 @@ public class StashUnstashPlugin implements Plugin, Resettable {
 
     public static withArtifactFrom(String patternFile) {
         this.patternFile = patternFile
+        return this
+    }
+
+    public static withUnstashVariable(String unstashVariable) {
+        this.unstashVariableName = unstashVariable
         return this
     }
 
@@ -22,8 +29,9 @@ public class StashUnstashPlugin implements Plugin, Resettable {
     }
 
     public static void init() {
-        StagePlugins.add(new StashUnstashPlugin(), BuildStage.class)
-        StagePlugins.add(new StashUnstashPlugin(), DeployStage.class)
+        def plugin = new StashUnstashPlugin()
+        StagePlugins.add(plugin, BuildStage.class)
+        StagePlugins.add(plugin, DeployStage.class)
     }
 
     public void apply(Stage stage) {
@@ -37,19 +45,31 @@ public class StashUnstashPlugin implements Plugin, Resettable {
     public Closure stashDecoration() {
         return { innerClosure ->
             innerClosure()
-            stash includes: getArtifactPattern(), name: DEFAULT_STASH_NAME
+            def artifactPattern = getArtifactPattern()
+            this.stashedFilename = sh(script: "ls ${artifactPattern}".toString(), returnStdout: true).trim()
+
+            stash includes: artifactPattern, name: DEFAULT_STASH_NAME
         }
+    }
+
+    public String getStashedFilename() {
+        return stashedFilename
+    }
+
+    public String getUnstashVariableName() {
+        return unstashVariableName ?: 'BUILD_ARTIFACT'
     }
 
     public Closure unstashDecoration() {
         return { innerClosure ->
             unstash DEFAULT_STASH_NAME
-            innerClosure()
+            withEnv(["${getUnstashVariableName()}=${getStashedFilename()}".toString()], innerClosure)
         }
     }
 
     public static reset() {
         pattern = null
         patternFile = null
+        unstashVariableName = null
     }
 }